requirements.txt
altair==4.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==22.2.0
azure-cognitiveservices-speech==1.19.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==4.1.0
blinker==1.5
cachetools==4.2.4
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
comm==0.1.4
commonmark==0.9.1
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
gitdb==4.0.9
GitPython==3.1.18
idna==3.6
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.8.1
jedi==0.17.2
Jinja2==3.0.3
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.7
MarkupSafe==2.0.1
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.8
notebook==6.4.10
numpy==1.19.5
openai==0.8.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.17.1
prompt-toolkit==3.0.19
protobuf==3.19.6
pyarrow==6.0.1
PyAudio @ file:///C:/AzureStudioChatGPTVoiceBot/PyAudio-0.2.11-cp36-cp36m-win_amd64.whl
pycparser==2.21
pydeck==0.6.2
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
pywin32==305
pywinpty==1.1.6
pyzmq==25.1.2
requests==2.27.1
rich==10.11.0
semver==2.13.0
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.0
SpeechRecognition==3.9.0
streamlit==1.10.0
streamlit-chat==0.1.1
terminado==0.12.1
testpath==0.6.0
toml==0.10.2
toolz==0.12.0
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typing-extensions==3.10.0.2
tzdata==2023.4
tzlocal==4.2
urllib3==1.26.18
validators==0.20.0
watchdog==2.3.1
wcwidth==0.2.12
webencodings==0.5.1
widgetsnbextension==3.6.6
zipp==3.6.0
